+++ lib/assets/core/javascripts/cartodb/organization/icon_picker/icon_picker_dialog/assets_template.jst.ejs
@@ -1,0 +1,2 @@
+<div class="FormAccount-row js-dialogIconPicker">
+</div>
