+++ lib/assets/core/javascripts/cartodb/organization/icon_picker/icon_picker_dialog/icon_picker_dialog_template.jst.ejs
@@ -1,7 +1,7 @@
 <div class="Modal-header">
   <div class="Modal-headerContainer">
     <h2 class="CDB-Text CDB-Size-huge is-light u-bSpace">Select marker image</h2>
-    <h3 class="CDB-Text CDB-Size-medium u-altTextColor"><a data-type="upload_file" href="#/upload" class="js-item">Upload your icons</a></h3>
+    <h3 class="CDB-Text CDB-Size-medium u-altTextColor"><a data-type="upload_file" href="#/upload" class="js-addIcon">Upload your icons</a></h3>
   </div>
 </div>
 
@@ -15,13 +15,6 @@
     <div class="Modal-inner">
       <div class="Dialog-body Dialog-body--expanded Dialog-body--create Dialog-body--withoutBorder">
         <div class="u-inner">
-          <div class="FormAccount-rowLabel">
-            <label class="CDB-Text CDB-Size-small u-lSpace--xl u-altTextColor js-runningInfo" style="display: none;"></label>
-            <label class="CDB-Text CDB-Size-medium u-lSpace--xl js-selectAllIcons" style="display: none;"><a href="#/select_all">Select all</a></label>
-            <label class="CDB-Text CDB-Size-medium u-lSpace--xl js-deselectAllIcons" style="display: none;"><a href="#/deselect_all">Deselect all</a></label>
-            <label class="CDB-Text CDB-Size-medium u-lSpace--xl js-deleteIcons" style="display: none;"><a class="IconText is--critical" href="#/delete_icons">Delete icons...</a></label>
-          </div>
-
           <div class="AssetsContent Dialog-bodyInnerExpandedWithSubFooter"></div>
         </div>
       </div>
