+++ lib/assets/core/javascripts/cartodb/organization/icon_picker/icon_picker_dialog/icon_picker_dialog_view.js
@@ -1,7 +1,7 @@
 var cdb = require('cartodb.js-v3');
 var ImagePickerDialog = require('../../../../../javascripts/cartodb/common/dialogs/map/image_picker_view');
 var ImagePickerNavigationView = require('./navigation_view');
-var AssetsView = require('./assets_view');
+// var AssetsView = require('./assets_view');
 
 // var UploadView = require('../../../../../javascripts/cartodb/common/dialogs/map/image_picker/file_upload_view');
 // var DropboxView = require('../../../../../javascripts/cartodb/common/dialogs/map/image_picker/dropbox_view');
@@ -15,14 +15,6 @@ module.exports = ImagePickerDialog.extend({
 
   className: "Dialog Modal ImagePicker",
 
-  events: {
-    'click .js-addIcon': '_onAddIconClicked',
-    'click .js-selectAllIcons': '_onSelectAllIconsClicked',
-    'click .js-deselectAllIcons': '_onDeselectAllIconsClicked',
-    'click .js-deleteIcons': '_onDeleteIconsClicked',
-    'change #iconfile': '_onFileSelected'
-  },
-
   initialize: function() {
     this.elder('initialize');
 
@@ -34,43 +26,9 @@ module.exports = ImagePickerDialog.extend({
       submit_enabled: false
     });
 
-    this.collection = this.options.collection;
-
     this._template = cdb.templates.getTemplate('organization/icon_picker/icon_picker_dialog/icon_picker_dialog_template');
     this._initBinds();
     this._onChangePane();
-  },
-
-  _onAddIconClicked: function () {
-    console.log('_onAddIconClicked');
-  },
-
-  _onSelectAllIconsClicked: function () {
-    console.log('_onSelectAllIconsClicked');
-  },
-
-  _onDeselectAllIconsClicked: function () {
-    console.log('_onDeselectAllIconsClicked');
-  },
-
-  _onDeleteIconsClicked: function () {
-    console.log('_onDeleteIconsClicked');
-  },
-
-  _onFileSelected: function () {
-    console.log('_onFileSelected');
-  },
-
-  _validate: function() {
-    if (!this.options.user) {
-      throw new TypeError('user is required');
-    }
-
-    if (!this.options.kind) {
-      throw new Error('kind should be passed');
-    }
-
-    if (!this.options.collection) throw new Error('collection is required');
   },
 
   _initViews: function() {
@@ -82,8 +40,6 @@ module.exports = ImagePickerDialog.extend({
     this._renderUploadLoader();
 
     this._renderUserIconsPane();
-    this._renderFilePane();
-    this._renderDropboxPane();
 
     this._contentPane.active('loader');
   },
@@ -113,10 +69,7 @@ module.exports = ImagePickerDialog.extend({
 
   _renderUserIconsPane: function() {
     var pane = new UserIconsView({
-      model: this.model,
-      collection: this.collection,
-      kind: this.kind,
-      folder: 'your-icons'
+      user: this.options.user
     });
 
     this._renderPane('your_icons', pane);
