+++ lib/assets/core/javascripts/cartodb/organization/icon_picker/icon_picker_dialog/user_icons_view.js
@@ -1,30 +1,30 @@
 var cdb = require('cartodb.js-v3');
-var _ = require('underscore-cdb-v3');
-var UserIconsView = require('../../../../../javascripts/cartodb/common/dialogs/map/image_picker/user_icons_view');
-var IconView = require('../icons/organization_icon_view');
+var DialogIconPickerView = require('../icons/organization_icons_view');
+var DialogIconCollection = require('../icons/organization_icon_collection');
 
-module.exports = UserIconsView.extend({
-
-  className: 'AssetPane AssetPane-userIcons',
+module.exports = cdb.core.View.extend({
 
   initialize: function() {
-    this.model = this.options.model;
-    this.template = cdb.templates.getTemplate('common/dialogs/map/image_picker/assets_template');
+    if (!this.options.user) { throw new Error('user ID is required.'); }
+    this._user = this.options.user;
 
-    this.collection.bind('add remove reset', this.render, this);
+    this.template = cdb.templates.getTemplate('organization/icon_picker/icon_picker_dialog/assets_template');
+    this.render();
+  },
+
+  render: function() {
+    this.clearSubViews();
+    this.$el.html(this.template());
+    this._renderAssets();
+    return this;
   },
 
   _renderAssets: function() {
-    var self = this;
-    var items = this.collection.where({ kind: this.options.kind });
-
-    _(items).each(function(iconModel) {
-      var iconView = new IconView({
-        model: iconModel
+    if (this.$('.js-dialogIconPicker').length > 0) {
+      this.icon_picker = new DialogIconPickerView({
+        el: this.$('.js-dialogIconPicker'),
+        user: this._user
       });
-      iconView.render();
-      self.$('ul').append(iconView.$el);
-      self.addView(iconView);
-    });
+    }
   }
 });
