+++ lib/assets/core/javascripts/cartodb/organization/icon_picker/icons/delete_icons_dialog_view.js
@@ -1,4 +1,4 @@
-var BaseDialog = require('../../common/views/base_dialog/view');
+var BaseDialog = require('../../../../../javascripts/cartodb/common/views/base_dialog/view');
 
 module.exports = BaseDialog.extend({
 
@@ -12,7 +12,7 @@ module.exports = BaseDialog.extend({
   },
 
   render_content: function() {
-    return this.getTemplate('organization/icons/delete_icons_modal')({
+    return this.getTemplate('organization/icon_picker/icons/delete_icons_modal')({
       numOfIcons: this._numOfIcons
     });
   },
