+++ lib/assets/core/javascripts/cartodb/organization/icon_picker/icons/delete_icons_modal.jst.ejs
