+++ lib/assets/core/javascripts/cartodb/organization/icon_picker/icons/organization_icons_template.jst.ejs
@@ -1,0 +1,18 @@
+<div class="FormAccount-rowLabel">
+  <label class="CDB-Text CDB-Size-medium is-semibold u-mainTextColor u-rSpace--xl js-iconMainLabel" style="display: none;"></label>
+  <label class="CDB-Text CDB-Size-small u-rSpace--xl u-altTextColor js-iconsInfo">These icons can be used as markers.</label>
+  <label class="CDB-Text CDB-Size-small u-rSpace--xl u-altTextColor js-runningInfo" style="display: none;"></label>
+  <label class="CDB-Text CDB-Size-medium u-rSpace--xl js-selectAllIcons" style="display: none;"><a href="#/select_all">Select all</a></label>
+  <label class="CDB-Text CDB-Size-medium u-rSpace--xl js-deselectAllIcons" style="display: none;"><a href="#/deselect_all">Deselect all</a></label>
+  <label class="CDB-Text CDB-Size-medium js-deleteIcons" style="display: none;"><a class="IconText is--critical" href="#/delete_icons">Delete icons...</a></label>
+</div>
+<div class="FormAccount-rowData js-asset-icons">
+  <ul class="IconList js-items">
+
+  </ul>
+</div>
+<div class="FormAccount-rowData js-errorMessage" style="display: none;">
+  <label class="CDB-Text CDB-Size-medium IconText--red"></label>
+  <label class="CDB-Text CDB-Size-medium IconText is--critical"></label>
+</div>
+<input type="file" value="Choose icon" accept="image/jpeg,image/jpg,image/gif,image/png,image/svg+xml" id="iconfile" tabindex="-1" style="position: absolute; clip: rect(0px 0px 0px 0px); opacity: 0;" multiple>
