+++ lib/assets/core/javascripts/cartodb/organization/icon_picker/organization_icon_picker_template.jst.ejs
@@ -1,11 +1,11 @@
 <div class="FormAccount-rowLabel">
-  <label class="CDB-Text CDB-Size-medium is-semibold u-mainTextColor js-iconMainLabel">Icons</label>
-  <label class="CDB-Text CDB-Size-medium u-lSpace--xl js-viewAllIcons" style="display: none;"><a href="#/view_icons">View all icons</a></label>
-  <label class="CDB-Text CDB-Size-small u-lSpace--xl u-altTextColor js-iconsInfo">These icons can be used as markers.</label>
-  <label class="CDB-Text CDB-Size-small u-lSpace--xl u-altTextColor js-runningInfo" style="display: none;"></label>
-  <label class="CDB-Text CDB-Size-medium u-lSpace--xl js-selectAllIcons" style="display: none;"><a href="#/select_all">Select all</a></label>
-  <label class="CDB-Text CDB-Size-medium u-lSpace--xl js-deselectAllIcons" style="display: none;"><a href="#/deselect_all">Deselect all</a></label>
-  <label class="CDB-Text CDB-Size-medium u-lSpace--xl js-deleteIcons" style="display: none;"><a class="IconText is--critical" href="#/delete_icons">Delete icons...</a></label>
+  <label class="CDB-Text CDB-Size-medium is-semibold u-mainTextColor u-rSpace--xl js-iconMainLabel">Icons</label>
+  <label class="CDB-Text CDB-Size-medium u-rSpace--xl js-viewAllIcons" style="display: none;"><a href="#/view_icons">View all icons</a></label>
+  <label class="CDB-Text CDB-Size-small u-rSpace--xl u-altTextColor js-iconsInfo">These icons can be used as markers.</label>
+  <label class="CDB-Text CDB-Size-small u-rSpace--xl u-altTextColor js-runningInfo" style="display: none;"></label>
+  <label class="CDB-Text CDB-Size-medium u-rSpace--xl js-selectAllIcons" style="display: none;"><a href="#/select_all">Select all</a></label>
+  <label class="CDB-Text CDB-Size-medium u-rSpace--xl js-deselectAllIcons" style="display: none;"><a href="#/deselect_all">Deselect all</a></label>
+  <label class="CDB-Text CDB-Size-medium js-deleteIcons" style="display: none;"><a class="IconText is--critical" href="#/delete_icons">Delete icons...</a></label>
 </div>
 <div class="FormAccount-rowData js-asset-icons">
   <ul class="IconList js-items">
